fix: 노선의 총 거리와 시간 자료형 변경

+++ modules/domain/src/main/kotlin/com/moyeobus/domain/route/Route.kt
@@ -10,9 +10,9 @@ data class Route (
 
     var busId: Long? = null,
 
-    val routeDistance: Double,
+    val routeDistance: Int,
 
-    val routeTotalTime: Double,
+    val routeTotalTime: Int,
 
     val routeComponents: List<RouteComponent> = emptyList()
 ) {
